Refactoring - Search Post / Hashtag
