+++ src/containers/HomePostList/index.tsx
@@ -1,0 +1,3 @@
+import HomePostList from "./HomePostList";
+
+export default HomePostList;
