--- src/containers/PostList/index.tsx
@@ -1,3 +1,0 @@
-import PostList from './PostList';
-
-export default PostList;
