+++ src/containers/SearchPostList/index.tsx
@@ -1,0 +1,3 @@
+import SerachPostList from "./SearchPostList";
+
+export default SerachPostList;
