--- src/pages/Search/index.tsx
@@ -1,3 +1,0 @@
-import Search from "./Search";
-
-export default Search;
