--- src/pages/SearchHashtag/index.tsx
@@ -1,3 +1,0 @@
-import SearchHashtag from "./SearchHashtag";
-
-export default SearchHashtag;
